Rename admin modules for clearer scope
- test-connection.ts → payment-test-buttons.ts: the module is
  specifically for Stripe/Square "Test Connection" buttons, not a
  generic connection tester
- question-visibility.ts → custom-question-visibility.ts: matches the
  fieldset.custom-question selector it targets
