+++ src/client/admin.ts
@@ -17,10 +17,10 @@ import { initManualCheckin } from "./admin/manual-checkin.ts";
 import { initMultiBookingBuilder } from "./admin/multi-booking.ts";
 import { initNavSelect } from "./admin/nav-select.ts";
 import { initPaymentResultNotifier } from "./admin/payment-result.ts";
-import { initQuestionVisibility } from "./admin/question-visibility.ts";
+import { initQuestionVisibility } from "./admin/custom-question-visibility.ts";
 import { initScrollHideNav } from "./admin/scroll-hide-nav.ts";
 import { initSelectOnClick } from "./admin/select-on-click.ts";
-import { initPaymentTestButtons } from "./admin/test-connection.ts";
+import { initPaymentTestButtons } from "./admin/payment-test-buttons.ts";
 
 initSelectOnClick();
 initNavSelect();
